@@ -35,7 +35,7 @@
     "sourcemap-lookup": "0.0.5",
     "testem": "^3.11.0",
     "uglify-js": "^3.17.4",
-    "webpack": "^5.89.0",
+    "webpack": "^5.90.3",
     "webpack-cli": "^5.1.4",
     "webpack-dev-middleware": "^7.0.0"
   },
